requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Machine Learning & Analytics
plotly>=5.0.0
statsmodels>=0.13.0

# AI & NLP Libraries
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
transformers>=4.20.0
torch>=1.12.0

# Document Processing
PyPDF2>=2.0.0
pdfplumber>=0.7.0
python-docx>=0.8.11

# Web & API (for future enhancements)
flask>=2.0.0
fastapi>=0.70.0
uvicorn>=0.15.0

# Data Visualization
bokeh>=2.4.0
altair>=4.2.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0
requests>=2.28.0

# Development & Testing
pytest>=7.0.0
jupyter>=1.0.0
ipython>=8.0.0

# Optional: For production deployment
gunicorn>=20.1.0
redis>=4.0.0